Updating to use Content Library

+++ deploy-amazon-linux2-hci-vcenter.ps1
@@ -2,14 +2,13 @@
 # www.cloudmethodsllc.com
 # Deployment of Amazon Linux 2 in vCenter/vSphere
 
+# Import PowerCLI Modules
+Import-Module VMware.VimAutomation.Core
+
+# Connect to VCenter (Prompt for user credentials)
 Connect-VIServer hci-vcenter.cetech-ne.local
 
-# Load OVF/OVA configuration from NAS share into a variable
-
-$ovfPath = "\\prodnas\downloads\OVAs\amzn2-vmware_esx-2.0.20210617.0-x86_64.xfs.gpt.ova"
-$ovfConfig = Get-OvfConfiguration -Ovf $ovfPath
-
-# vSphere Cluster + Network configurations
+# vSphere Cluster + Network configuration parameters
 $Cluster = Get-Cluster -Name "NTAP"
 $VMName = "cetech-amzn2"
 $VMNetwork = "VM_Network"
@@ -22,7 +21,12 @@ $DiskFormat = "Thin"
 $Folder = "Templates"
 
 # Delete existing template
-Remove-Template -Template $VMName -DeletePermanently -Confirm:$false
+$template = Get-Template $VMName -ErrorAction SilentlyContinue
+if ($template) {
+    Remove-Template -Template $VMName -DeletePermanently -Confirm:$false
+} else {
+    Write-Host "No existing template found"
+}
 
 # Copy seed.iso file from NAS share to datastore
 $DatastoreTemp = Get-Datastore $DatastoreISO
@@ -30,13 +34,17 @@ New-PSDrive -Location $DatastoreTemp -Name ds -PSProvider VimDatastore -Root "\"
 Copy-DatastoreItem -Item "\\prodnas\downloads\OVAs\seed.iso" -Destination "ds:\ISO\Amazon\"
 Remove-PSDrive -Name ds
 
-# vSphere Network Mapping based on OVF/OVA configuration
+# Fetch OVA from Content Library
+$ova = Get-ContentLibraryItem -ContentLibrary cetech-images -Name cetech-amzn2
+
+# Build OVF Configuration for OVA
+$userData = Get-Content -Path '.\user-data' -Raw
+$ovfConfig = Get-OvfConfiguration -ContentLibraryItem $ova -Target $VMHost
+# $ovfConfig.Common.user_data.Value = [System.Convert]::ToBase64String([System.Text.Encoding]::UTF8.GetBytes($userData))
 $ovfConfig.NetworkMapping.bridged.Value = $VMNetwork
 
-# Deploy the OVF/OVA with the config parameters
-Import-VApp -Source $ovfpath -OvfConfiguration $ovfConfig -Name $VMName -VMHost $VMHost -Location $Cluster -Datastore $Datastore -DiskStorageFormat $DiskFormat -InventoryLocation $Folder -Confirm:$false
-
-$VM = Get-VM $VMName
+# Launch VM from OVA
+$VM = New-VM -ContentLibraryItem $ova -Name $VMName -VMHost $VMHost -Location $Folder -Datastore $Datastore -DiskStorageFormat $DiskFormat -Confirm:$false
 
 # Add CD-Drive to VM and mount seed.iso
 New-CDDrive -VM $VM -IsoPath "[$DatastoreISO] ISO\Amazon\seed.iso" -StartConnected
@@ -54,8 +62,11 @@ Shutdown-VMGuest $VM -Confirm:$false
 Start-Sleep -Seconds 30
 
 # Assign CD to variable in order to remove it
-$CD = Get-CDDrive -VM $VM
-Remove-CDDrive -CD $CD -Confirm:$false
+### $CD = Get-CDDrive -VM $VM
+### Remove-CDDrive -CD $CD -Confirm:$false
 
 # Convert VM to Template
 Get-VM -Name $VMName | Set-VM -ToTemplate -Confirm:$false
+
+# Disconnect from VCenter
+Disconnect-VIServer -Confirm:$false
